requirements.txt
streamlit==1.28.2
tensorflow==2.17.0
Pillow==9.4.0
numpy==1.26.1